protocol base for subtraction and adjust
