--- xmipp3/protocols/protocol_volume_adjust.py
@@ -1,134 +1,0 @@
-# -*- coding: utf-8 -*-
-# **************************************************************************
-# *
-# * Authors:  Estrella Fernandez Gimenez (me.fernandez@cnb.csic.es)
-# *
-# *
-# * Unidad de  Bioinformatica of Centro Nacional de Biotecnologia , CSIC
-# *
-# * This program is free software; you can redistribute it and/or modify
-# * it under the terms of the GNU General Public License as published by
-# * the Free Software Foundation; either version 2 of the License, or
-# * (at your option) any later version.
-# *
-# * This program is distributed in the hope that it will be useful,
-# * but WITHOUT ANY WARRANTY; without even the implied warranty of
-# * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# * GNU General Public License for more details.
-# *
-# * You should have received a copy of the GNU General Public License
-# * along with this program; if not, write to the Free Software
-# * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA
-# * 02111-1307  USA
-# *
-# *  All comments concerning this program package may be sent to the
-# *  e-mail address 'scipion@cnb.csic.es'
-# *
-# **************************************************************************
-
-from pyworkflow.protocol.params import PointerParam, BooleanParam, IntParam, FloatParam, EnumParam, StringParam, FileParam
-from pyworkflow.protocol.constants import LEVEL_ADVANCED
-
-from pwem.convert import headers, downloadPdb, cifToPdb
-from pwem.objects import Volume, Transform
-from pwem.protocols import EMProtocol
-
-
-class XmippProtVolAdjust(EMProtocol):
-    """ This protocol scales a volume in order to assimilate it to another one.
-    The volume with the best resolution should be the first one.
-    The volumes should be aligned previously and they have to be equal in size"""
-
-    _label = 'volumes adjust'
-    IMPORT_OBJ = 0
-    IMPORT_FROM_FILES = 1
-
-    # --------------------------- DEFINE param functions --------------------------------------------
-    def _defineParams(self, form):
-
-        form.addSection(label='Input')
-        form.addParam('vol1', PointerParam, pointerClass='Volume', label="Volume 1 ", help='Specify a volume.')
-        form.addParam('vol2', PointerParam, pointerClass='Volume', label="Volume 2 ", help='Specify a volume.')
-        form.addParam('masks', BooleanParam, label='Mask volumes?', default=True,
-                      help='The masks are not mandatory but highly recommendable.')
-        form.addParam('mask1', PointerParam, pointerClass='VolumeMask', label="Mask for volume 1",
-                      condition='masks', help='Specify a mask for volume 1.')
-        form.addParam('mask2', PointerParam, pointerClass='VolumeMask', label="Mask for volume 2",
-                      condition='masks', help='Specify a mask for volume 1.')
-        form.addParam('resol', FloatParam, label="Resolution of volume 2: ",
-                      help='Resolution of volume 2 for filtering. ')
-        form.addParam('sigma', FloatParam, label="Decay of the filter (sigma): ", default=3, condition='resol',
-                      help='Decay of the filter (sigma parameter) to smooth the mask transition',
-                      expertLevel=LEVEL_ADVANCED)
-        form.addParam('iter', IntParam, label="Number of iterations: ", default=5, expertLevel=LEVEL_ADVANCED)
-        form.addParam('rfactor', FloatParam, label="Relaxation factor (lambda): ", default=0.5,
-                      expertLevel=LEVEL_ADVANCED,
-                      help='Relaxation factor for Fourier amplitude projector (POCS), it should be between 0 and 1, '
-                           'being 0 no relaxation and 1 no modification of volume 2 amplitudes')
-
-    # --------------------------- INSERT steps functions --------------------------------------------
-    def _insertAllSteps(self):
-        self._insertFunctionStep('adjustStep')
-        self._insertFunctionStep('createOutputStep')
-
-    # --------------------------- STEPS functions --------------------------------------------
-    def adjustStep(self):
-        vol1 = self.vol1.get().clone()
-        vol2 = self.vol2.get().getFileName()
-        if self.masks:
-            mask2 = self.mask2.get().getFileName()
-        resol = self.resol.get()
-        iter = self.iter.get()
-        program = "xmipp_volume_subtraction"
-        args = '--i1 %s --i2 %s -o %s --iter %s --lambda %s' % \
-               (vol1.getFileName(), vol2, self._getExtraPath("output_volume.mrc"), iter, self.rfactor.get())
-        if resol:
-            fc = vol1.getSamplingRate()/resol
-            args += ' --cutFreq %f --sigma %d' % (fc, self.sigma.get())
-        if self.masks:
-            args += ' --mask1 %s --mask2 %s' % (self.mask1.get().getFileName(), mask2)
-        self.runJob(program, args)
-
-    def createOutputStep(self):
-        vol1 = self.vol1.get()
-        volume = Volume()
-        volume.setSamplingRate(vol1.getSamplingRate())
-        if vol1.getFileName().endswith('mrc'):
-            origin = Transform()
-            ccp4header = headers.Ccp4Header(vol1.getFileName(), readHeader=True)
-            shifts = ccp4header.getOrigin()
-            origin.setShiftsTuple(shifts)
-            volume.setOrigin(origin)
-        volume.setFileName(self._getExtraPath("output_volume.mrc"))
-        filename = volume.getFileName()
-        if filename.endswith('.mrc') or filename.endswith('.map'):
-            volume.setFileName(filename + ':mrc')
-        self._defineOutputs(outputVolume=volume)
-
-    # --------------------------- INFO functions --------------------------------------------
-    def _summary(self):
-        summary = ["Volume 1: %s\nVolume 2: %s" % (self.vol1.get().getFileName(), self.vol2.get().getFileName())]
-        if self.masks:
-            summary.append("Input mask 1: %s" % self.mask1.get().getFileName())
-            summary.append("Input mask 2: %s" % self.mask2.get().getFileName())
-        if self.resol.get() != 0:
-            summary.append("Filter at resolution %f A" % self.resol.get())
-        return summary
-
-    def _methods(self):
-        methods = []
-        if not hasattr(self, 'outputVolume'):
-            methods.append("Output volume not ready yet.")
-        else:
-            methods.append("Volume %s adjusted to volume %s" % (self.vol2.get().getFileName(),
-                                                                self.vol1.get().getFileName()))
-            if self.resol.get() != 0:
-                methods.append(" at resolution %f A" % self.resol.get())
-        return methods
-
-    def _validate(self):
-        errors = []
-        rfactor = self.rfactor.get()
-        if rfactor < 0 or rfactor > 1:
-            errors.append('Relaxation factor (lambda) must be between 0 and 1')
-        return errors
