+++ xmipp3/protocols/protocol_volume_adjust_sub.py
@@ -35,7 +35,56 @@ from pwem.objects import Volume, Transform
 from pwem.protocols import EMProtocol
 
 
-class XmippProtVolSubtraction(EMProtocol):
+class XmippProtVolAdjBase(EMProtocol):
+    """ Helper class that contains some Protocol utilities methods
+    used by both  XmippProtVolSubtraction and XmippProtVolAdjust."""
+
+    # --------------------------- DEFINE param functions --------------------------------------------
+    @classmethod
+    def _defineParams(self, form):
+        form.addSection(label='Input')
+        form.addParam('vol1', PointerParam, pointerClass='Volume', label="Volume 1 ", help='Specify a volume.')
+        form.addParam('masks', BooleanParam, label='Mask volumes?', default=True,
+                      help='The masks are not mandatory but highly recommendable.')
+        form.addParam('mask1', PointerParam, pointerClass='VolumeMask', label="Mask for volume 1",
+                      condition='masks', help='Specify a mask for volume 1.')
+        form.addParam('resol', FloatParam, label="Filter at resolution: ", default=3, allowsNull=True,
+                      expertLevel=LEVEL_ADVANCED,
+                      help='Resolution (A) at which subtraction will be performed, filtering the input volumes.'
+                           'Value 0 implies no filtering.')
+        form.addParam('sigma', FloatParam, label="Decay of the filter (sigma): ", default=3, condition='resol',
+                      help='Decay of the filter (sigma parameter) to smooth the mask transition',
+                      expertLevel=LEVEL_ADVANCED)
+        form.addParam('iter', IntParam, label="Number of iterations: ", default=5, expertLevel=LEVEL_ADVANCED)
+        form.addParam('rfactor', FloatParam, label="Relaxation factor (lambda): ", default=0.5,
+                      expertLevel=LEVEL_ADVANCED,
+                      help='Relaxation factor for Fourier amplitude projector (POCS), it should be between 0 and 1, '
+                           'being 0 no relaxation and 1 no modification of volume 2 amplitudes')
+
+    # --------------------------- INSERT steps functions --------------------------------------------
+    def _insertAllSteps(self):
+        self._insertAdjSteps()
+        self._insertFunctionStep('createOutputStep')
+
+    # --------------------------- STEPS functions --------------------------------------------
+    def createOutputStep(self):
+        vol1 = self.vol1.get()
+        volume = Volume()
+        volume.setSamplingRate(vol1.getSamplingRate())
+        if vol1.getFileName().endswith('mrc'):
+            origin = Transform()
+            ccp4header = headers.Ccp4Header(vol1.getFileName(), readHeader=True)
+            shifts = ccp4header.getOrigin()
+            origin.setShiftsTuple(shifts)
+            volume.setOrigin(origin)
+        volume.setFileName(self._getExtraPath("output_volume.mrc"))
+        filename = volume.getFileName()
+        if filename.endswith('.mrc') or filename.endswith('.map'):
+            volume.setFileName(filename + ':mrc')
+        self._defineOutputs(outputVolume=volume)
+
+
+class XmippProtVolSubtraction(XmippProtVolAdjBase):
     """ This protocol scales a volume in order to adjust it to another one. Then, it can calculate the subtraction
     of the two volumes. Second input can be a pdb. The volumes should be aligned previously and they have to
     be equal in size"""
@@ -46,9 +95,7 @@ class XmippProtVolSubtraction(EMProtocol):
 
     # --------------------------- DEFINE param functions --------------------------------------------
     def _defineParams(self, form):
-
-        form.addSection(label='Input')
-        form.addParam('vol1', PointerParam, pointerClass='Volume', label="Volume 1 ", help='Specify a volume.')
+        XmippProtVolAdjBase._defineParams(form)
         form.addParam('pdb', BooleanParam, label='Is the second input a PDB?', default=False,
                       help='If yes, the protocol will generate and store in folder "extra" of this protocol '
                            'a volume and a mask from the pdb. If not, a second volume has to be input and optionally '
@@ -65,31 +112,15 @@ class XmippProtVolSubtraction(EMProtocol):
                       help='Specify a path to desired PDB structure.')
         form.addParam('vol2', PointerParam, pointerClass='Volume', label="Volume 2 ", condition='pdb == False',
                       help='Specify a volume.')
-        form.addParam('masks', BooleanParam, label='Mask volumes?', default=True,
-                      help='The masks are not mandatory but highly recommendable.')
-        form.addParam('mask1', PointerParam, pointerClass='VolumeMask', label="Mask for volume 1",
-                      condition='masks', help='Specify a mask for volume 1.')
         form.addParam('mask2', PointerParam, pointerClass='VolumeMask', label="Mask for volume 2",
                       condition='masks and pdb==False', help='Specify a mask for volume 1.')
-        form.addParam('resol', FloatParam, label="Subtraction at resolution: ", default=3, allowsNull=True,
-                      help='Resolution (A) at which subtraction will be performed, filtering the input volumes.'
-                           'Value 0 implies no filtering.')
-        form.addParam('sigma', FloatParam, label="Decay of the filter (sigma): ", default=3, condition='resol',
-                      help='Decay of the filter (sigma parameter) to smooth the mask transition',
-                      expertLevel=LEVEL_ADVANCED)
-        form.addParam('iter', IntParam, label="Number of iterations: ", default=5, expertLevel=LEVEL_ADVANCED)
-        form.addParam('rfactor', FloatParam, label="Relaxation factor (lambda): ", default=0.5,
-                      expertLevel=LEVEL_ADVANCED,
-                      help='Relaxation factor for Fourier amplitude projector (POCS), it should be between 0 and 1, '
-                           'being 0 no relaxation and 1 no modification of volume 2 amplitudes')
 
     # --------------------------- INSERT steps functions --------------------------------------------
-    def _insertAllSteps(self):
+    def _insertAdjSteps(self):
         if self.pdb:
             self._insertFunctionStep('convertPdbStep')
             self._insertFunctionStep('generateMask2Step')
         self._insertFunctionStep('subtractionStep')
-        self._insertFunctionStep('createOutputStep')
 
     # --------------------------- STEPS functions --------------------------------------------
     def convertPdbStep(self):
@@ -155,22 +186,6 @@ class XmippProtVolSubtraction(EMProtocol):
             args += ' --mask1 %s --mask2 %s' % (self.mask1.get().getFileName(), mask2)
         self.runJob(program, args)
 
-    def createOutputStep(self):
-        vol1 = self.vol1.get()
-        volume = Volume()
-        volume.setSamplingRate(vol1.getSamplingRate())
-        if vol1.getFileName().endswith('mrc'):
-            origin = Transform()
-            ccp4header = headers.Ccp4Header(vol1.getFileName(), readHeader=True)
-            shifts = ccp4header.getOrigin()
-            origin.setShiftsTuple(shifts)
-            volume.setOrigin(origin)
-        volume.setFileName(self._getExtraPath("output_volume.mrc"))
-        filename = volume.getFileName()
-        if filename.endswith('.mrc') or filename.endswith('.map'):
-            volume.setFileName(filename + ':mrc')
-        self._defineOutputs(outputVolume=volume)
-
     # --------------------------- INFO functions --------------------------------------------
     def _summary(self):
         summary = ["Volume 1: %s" % self.vol1.get().getFileName()]
@@ -206,6 +221,7 @@ class XmippProtVolSubtraction(EMProtocol):
         if rfactor < 0 or rfactor > 1:
             errors.append('Relaxation factor (lambda) must be between 0 and 1')
         return errors
+
     # --------------------------- UTLIS functions --------------------------------------------
     def _getPdbFileName(self):
         if self.inputPdbData == self.IMPORT_OBJ:
@@ -215,3 +231,68 @@ class XmippProtVolSubtraction(EMProtocol):
 
     def _getVolName(self):
         return self._getExtraPath(replaceBaseExt(self._getPdbFileName(), "vol"))
+
+
+class XmippProtVolAdjust(XmippProtVolAdjBase):
+    """ This protocol scales a volume in order to assimilate it to another one.
+    The volume with the best resolution should be the first one.
+    The volumes should be aligned previously and they have to be equal in size"""
+
+    _label = 'volumes adjust'
+    IMPORT_OBJ = 0
+    IMPORT_FROM_FILES = 1
+
+    # --------------------------- DEFINE param functions --------------------------------------------
+    def _defineParams(self, form):
+        XmippProtVolAdjBase._defineParams(form)
+        form.addParam('vol2', PointerParam, pointerClass='Volume', label="Volume 2 ", help='Specify a volume.')
+        form.addParam('mask2', PointerParam, pointerClass='VolumeMask', label="Mask for volume 2",
+                      condition='masks', help='Specify a mask for volume 1.')
+
+    # --------------------------- INSERT steps functions --------------------------------------------
+    def _insertAdjSteps(self):
+        self._insertFunctionStep('adjustStep')
+
+    # --------------------------- STEPS functions --------------------------------------------
+    def adjustStep(self):
+        vol1 = self.vol1.get().clone()
+        vol2 = self.vol2.get().getFileName()
+        resol = self.resol.get()
+        iter = self.iter.get()
+        program = "xmipp_volume_subtraction"
+        args = '--i1 %s --i2 %s -o %s --iter %s --lambda %s' % \
+               (vol1.getFileName(), vol2, self._getExtraPath("output_volume.mrc"), iter, self.rfactor.get())
+        if resol:
+            fc = vol1.getSamplingRate()/resol
+            args += ' --cutFreq %f --sigma %d' % (fc, self.sigma.get())
+        if self.masks:
+            args += ' --mask1 %s --mask2 %s' % (self.mask1.get().getFileName(), self.mask2.get().getFileName())
+        self.runJob(program, args)
+
+    # --------------------------- INFO functions --------------------------------------------
+    def _summary(self):
+        summary = ["Volume 1: %s\nVolume 2: %s" % (self.vol1.get().getFileName(), self.vol2.get().getFileName())]
+        if self.masks:
+            summary.append("Input mask 1: %s" % self.mask1.get().getFileName())
+            summary.append("Input mask 2: %s" % self.mask2.get().getFileName())
+        if self.resol.get() != 0:
+            summary.append("Filter at resolution %f A" % self.resol.get())
+        return summary
+
+    def _methods(self):
+        methods = []
+        if not hasattr(self, 'outputVolume'):
+            methods.append("Output volume not ready yet.")
+        else:
+            methods.append("Volume %s adjusted to volume %s" % (self.vol2.get().getFileName(),
+                                                                self.vol1.get().getFileName()))
+            if self.resol.get() != 0:
+                methods.append(" at resolution %f A" % self.resol.get())
+        return methods
+
+    def _validate(self):
+        errors = []
+        rfactor = self.rfactor.get()
+        if rfactor < 0 or rfactor > 1:
+            errors.append('Relaxation factor (lambda) must be between 0 and 1')
+        return errors
